chore: refactor to fix file too long
